Add grid functions.
Add and remove gem. both take a place which can be a Grid::Point or an
index.
Unused process_click().
new_waiting_gem() adds a gem above the game board ready for dropping.
to_index() examiness its parameter and only converts it if it's a
Grid::Point to begin with.
valid?() now allows for the y co-ordinate to be -1, i.e. the waiting
gem.